`episode_type` and `case_for_recurrence` at `episode_level`
Length of `episode_type`  can now be the same as `date`. This allows for simultaneous tracking of `"fixed"` and `"rolling"` episodes.

Length of `case_for_recurrence`  can now be the same as `date`. This allows for simultaneous tracking of `"rolling"` episodes with and without an additional case window for recurrent events.